fix: Handles to unauthenticated cases.
When accessing directly a meeting link from bookmark, make sure we redirect to login page.
When requesting meeting link without authentication from the page - clicking get link, make sure fetch does not redirect, but we handle it manually and redirect the whole page.
Make sure we remove the jwt from url so people do not bookmark the jwt which will expire in an hour.